Testing weird issue

+++ psdtypes.h
@@ -16,9 +16,9 @@ typedef struct{
 }psdtypesABC_t;
 
 typedef struct{
-	int32_t a;
-	int32_t b;
-	int32_t c;
+	int64_t a;
+	int64_t b;
+	int64_t c;
 }psdtypesABCint_t;
 
 typedef struct{
@@ -28,10 +28,28 @@ typedef struct{
 }psdtypesDQ0_t;
 
 typedef struct{
-	int32_t d;
-	int32_t q;
-	int32_t z;
+	int64_t d;
+	int64_t q;
+	int64_t z;
 }psdtypesDQ0int_t;
+
+//typedef struct{
+//	int32_t a;
+//	int32_t b;
+//	int32_t c;
+//}psdtypesABCint_t;
+//
+//typedef struct{
+//	float d;
+//	float q;
+//	float z;
+//}psdtypesDQ0_t;
+//
+//typedef struct{
+//	int32_t d;
+//	int32_t q;
+//	int32_t z;
+//}psdtypesDQ0int_t;
 
 typedef struct{
 	psdtypesDQ0_t ii;
