+++ tlvsi.h
@@ -63,13 +63,13 @@ void tlvsiPredict(psdtypesDQ0_t *ii_k_1, psdtypesDQ0_t *ii_k,
 				  psdtypesDQ0_t *vg_k,
 				  float theta, uint32_t sw);
 
-uint32_t tlvsiOptFixed(psdtypesDQ0int_t *ii, psdtypesDQ0int_t *ig, psdtypesDQ0int_t *vc, psdtypesDQ0int_t *vg, float theta, int32_t *Jopt);
+uint32_t tlvsiOptFixed(psdtypesDQ0int_t *ii, psdtypesDQ0int_t *ig, psdtypesDQ0int_t *vc, psdtypesDQ0int_t *vg, float theta, int64_t *Jopt);
 
 float tlvsiCost(psdtypesDQ0_t *ii, psdtypesDQ0_t *ii_ref,
 				psdtypesDQ0_t *ig, psdtypesDQ0_t *ig_ref,
 				psdtypesDQ0_t *vc, psdtypesDQ0_t *vc_ref);
 
-int32_t tlvsiCostFixed(psdtypesDQ0int_t *ii, psdtypesDQ0int_t *ii_ref,
+int64_t tlvsiCostFixed(psdtypesDQ0int_t *ii, psdtypesDQ0int_t *ii_ref,
 					 psdtypesDQ0int_t *ig, psdtypesDQ0int_t *ig_ref,
 					 psdtypesDQ0int_t *vc, psdtypesDQ0int_t *vc_ref);
 //===========================================================================
